+++ project/src/types/favorite-data.ts
@@ -3,5 +3,5 @@ import {CardTypes, FavoriteStatus} from '../const';
 export type FavoriteData = {
   id: number,
   status: FavoriteStatus,
-  page: CardTypes,
+  cardType: CardTypes,
 };
